requirements.txt
django
djangorestframework
djangorestframework.simplejwt
psycopg2-binary
django-heroku==0.3.1
gunicorn==20.1.0